requirements.txt
transformers
google-cloud-aiplatform
tiktoken
unstructured
tqdm
tabulate
pytesseract
pdf2image
datasets
accelerate
google-generativeai
scikit-learn
rouge_score 
nltk
deepspeed
tensorboard